push

+++ fdf/src/utils_3d/structor.c
@@ -1,6 +1,6 @@
 #include "vector3d.h"
 
-t_vector3d	vector3d(double x, double y, double z)
+t_vector3d	vector3d(float x, float y, float z)
 {
 	return ((t_vector3d){x, y, z});
 }
